+++ tools/ensemble_test_runner/CHANGELOG.md
@@ -1,10 +1,18 @@
+## 1.2.52
+
+- feat(ensemble_test_runner): implement mock asset path resolution and add tests for nested mocks
+
+## 1.2.51
+
+- fix dependency issues
+
 ## 1.2.50
 
- - **FEAT**(test_runner): support live Firebase and HTTP under flutter test ([#2321](https://github.com/ensembleUI/ensemble/issues/2321)). ([9ffe6994](https://github.com/ensembleUI/ensemble/commit/9ffe6994322370ca32a79f1ce75eb06bee251018))
+- **FEAT**(test_runner): support live Firebase and HTTP under flutter test ([#2321](https://github.com/ensembleUI/ensemble/issues/2321)). ([9ffe6994](https://github.com/ensembleUI/ensemble/commit/9ffe6994322370ca32a79f1ce75eb06bee251018))
 
 ## 1.2.49
 
- - Bump "ensemble_test_runner" to `1.2.49`.
+- Bump "ensemble_test_runner" to `1.2.49`.
 
 # Changelog